refactor: async iterators wip
