+++ package.json
@@ -42,13 +42,10 @@
   },
   "dependencies": {
     "abort-controller": "^3.0.0",
-    "async-iterator-all": "^1.0.0",
-    "async-iterator-to-pull-stream": "^1.3.0",
     "bignumber.js": "^9.0.0",
     "bl": "^4.0.0",
     "bs58": "^4.0.1",
     "buffer": "^5.4.2",
-    "callbackify": "^1.1.0",
     "cids": "~0.7.1",
     "debug": "^4.1.0",
     "err-code": "^2.0.0",
@@ -62,7 +59,6 @@
     "is-ipfs": "~0.6.1",
     "it-glob": "0.0.6",
     "it-tar": "^1.1.0",
-    "it-to-stream": "^0.1.1",
     "iterable-ndjson": "^1.1.0",
     "ky": "^0.15.0",
     "ky-universal": "^0.3.0",
@@ -74,8 +70,7 @@
     "multihashes": "~0.4.14",
     "parse-duration": "^0.1.1",
     "peer-id": "~0.12.3",
-    "peer-info": "~0.15.1",
-    "promise-nodeify": "^3.0.1"
+    "peer-info": "~0.15.1"
   },
   "devDependencies": {
     "aegir": "^20.4.1",
