+++ src/block/rm.js
