+++ src/bootstrap/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  add: callbackify.variadic(require('./add')(config)),
-  rm: callbackify.variadic(require('./rm')(config)),
-  list: callbackify.variadic(require('./list')(config))
+  add: require('./add')(config),
+  rm: require('./rm')(config),
+  list: require('./list')(config)
 })
