+++ src/config/index.js
@@ -1,10 +1,8 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  get: callbackify.variadic(require('./get')(config)),
-  set: callbackify.variadic(require('./set')(config)),
-  replace: callbackify.variadic(require('./replace')(config)),
+  get: require('./get')(config),
+  set: require('./set')(config),
+  replace: require('./replace')(config),
   profiles: require('./profiles')(config)
 })
