+++ src/config/profiles/index.js
@@ -1,8 +1,6 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  apply: callbackify.variadic(require('./apply')(config)),
-  list: callbackify.variadic(require('./list')(config))
+  apply: require('./apply')(config),
+  list: require('./list')(config)
 })
