+++ src/dag/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  get: callbackify.variadic(require('./get')(config)),
-  put: callbackify.variadic(require('./put')(config)),
-  resolve: callbackify.variadic(require('./resolve')(config))
+  get: require('./get')(config),
+  put: require('./put')(config),
+  resolve: require('./resolve')(config)
 })
