+++ src/diag/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  net: callbackify.variadic(require('./net')(config)),
-  sys: callbackify.variadic(require('./sys')(config)),
-  cmds: callbackify.variadic(require('./cmds')(config))
+  net: require('./net')(config),
+  sys: require('./sys')(config),
+  cmds: require('./cmds')(config)
 })
