+++ src/key/index.js
@@ -1,12 +1,10 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  gen: callbackify.variadic(require('./gen')(config)),
-  list: callbackify.variadic(require('./list')(config)),
-  rename: callbackify.variadic(require('./rename')(config)),
-  rm: callbackify.variadic(require('./rm')(config)),
-  export: callbackify.variadic(require('./export')(config)),
-  import: callbackify.variadic(require('./import')(config))
+  gen: require('./gen')(config),
+  list: require('./list')(config),
+  rename: require('./rename')(config),
+  rm: require('./rm')(config),
+  export: require('./export')(config),
+  import: require('./import')(config)
 })
