+++ src/log/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
   tail: require('./tail')(config),
-  ls: callbackify.variadic(require('./ls')(config)),
-  level: callbackify.variadic(require('./level')(config))
+  ls: require('./ls')(config),
+  level: require('./level')(config)
 })
