+++ src/name/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  publish: callbackify.variadic(require('./publish')(config)),
-  resolve: callbackify.variadic(require('./resolve')(config)),
+  publish: require('./publish')(config),
+  resolve: require('./resolve')(config),
   pubsub: require('./pubsub')(config)
 })
