+++ src/name/pubsub/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  cancel: callbackify.variadic(require('./cancel')(config)),
-  state: callbackify.variadic(require('./state')(config)),
-  subs: callbackify.variadic(require('./subs')(config))
+  cancel: require('./cancel')(config),
+  state: require('./state')(config),
+  subs: require('./subs')(config)
 })
