+++ src/object/patch/index.js
@@ -1,10 +1,8 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  addLink: callbackify.variadic(require('./add-link')(config)),
-  appendData: callbackify.variadic(require('./append-data')(config)),
-  rmLink: callbackify.variadic(require('./rm-link')(config)),
-  setData: callbackify.variadic(require('./set-data')(config))
+  addLink: require('./add-link')(config),
+  appendData: require('./append-data')(config),
+  rmLink: require('./rm-link')(config),
+  setData: require('./set-data')(config)
 })
