+++ src/pin/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-
 module.exports = config => ({
-  add: callbackify.variadic(require('./add')(config)),
-  rm: callbackify.variadic(require('./rm')(config)),
-  ls: callbackify.variadic(require('./ls')(config))
+  add: require('./add')(config),
+  rm: require('./rm')(config),
+  ls: require('./ls')(config)
 })
