+++ src/repo/index.js
@@ -1,10 +1,7 @@
 'use strict'
 
-const callbackify = require('callbackify')
-const { collectify } = require('../lib/converters')
-
 module.exports = config => ({
-  gc: callbackify.variadic(collectify(require('./gc')(config))),
-  stat: callbackify.variadic(require('./stat')(config)),
-  version: callbackify.variadic(require('./version')(config))
+  gc: require('./gc')(config),
+  stat: require('./stat')(config),
+  version: require('./version')(config)
 })
